+++ shaka-player.uncompiled.js
@@ -46,8 +46,9 @@ goog.require('shaka.offline.OfflineScheme');
 goog.require('shaka.offline.Storage');
 goog.require('shaka.offline.indexeddb.StorageMechanism');
 goog.require('shaka.polyfill.Aria');
-goog.require('shaka.polyfill.EncryptionScheme');
+goog.require('shaka.polyfill.EmeEncryptionScheme');
 goog.require('shaka.polyfill.Fullscreen');
+goog.require('shaka.polyfill.MCapEncryptionScheme');
 goog.require('shaka.polyfill.MediaSource');
 goog.require('shaka.polyfill.MediaCapabilities');
 goog.require('shaka.polyfill.Orientation');
